chore: Refactor code to move dns & proxmox into a package
This moves proxmox and DNS configs into a package. This makes for
a better Golang organization.
